requirements.txt
asgiref==3.2.3
Django==3.0.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
mongoengine==0.19.1
djongo
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.8.4
PyJWT==1.7.1
six==1.12.0
gunicorn==19.9.0
django-heroku
dnspython==1.16.0
